refactor: extract signal classifier into ai package
