+++ ai/src/__mocks__/server-only.ts
@@ -1,0 +1,1 @@
+export {};
