+++ ai/src/signal-classifier/schema.ts
@@ -1,0 +1,8 @@
+import { signalClassificationSchema } from "@repo/api-contract";
+
+// `schemaVersion` is a fixed, code-owned literal. The model owns only the
+// classification fields; runtime code stamps the schema version after parsing.
+export const signalClassificationModelSchema =
+  signalClassificationSchema.omit({
+    schemaVersion: true,
+  });
