+++ ai/vitest.config.ts
@@ -1,0 +1,19 @@
+import sharedConfig from "@repo/vitest-config";
+import { defineConfig, mergeConfig } from "vitest/config";
+
+export default mergeConfig(
+  sharedConfig,
+  defineConfig({
+    resolve: {
+      alias: {
+        "server-only": new URL(
+          "./src/__mocks__/server-only.ts",
+          import.meta.url
+        ).pathname,
+      },
+    },
+    test: {
+      environment: "node",
+    },
+  })
+);
